moved out of minirig
