+++ pacman/operations/router_compressors/mundys_router_compressor/exceptions.py
@@ -34,18 +34,3 @@ class MinimisationFailedError(Exception):
             text += " Best managed was {} entries.".format(self.final_length)
 
         return text
-
-
-class MultisourceRouteError(Exception):
-    """Indicates that two nets with the same key and mask would cause packets
-    to become duplicated.
-    """
-    def __init__(self, key, mask, coordinate):
-        self.key = key
-        self.mask = mask
-        self.x, self.y = coordinate
-
-    def __str__(self):
-        return ("Two different nets with the same key ({0.key:#010x}) and "
-                "mask ({0.mask:#010x}) fork differently at ({0.x}, {0.y})".
-                format(self))
